Update texteditor while source mode
CU-8694r63fj
