--- nextgisweb/gui/nodepkg/component/text-editor/type.ts
@@ -1,6 +1,0 @@
-export interface TextEditorProps {
-    value: string;
-    onChange?: (val: string) => void;
-    parentHeight?: boolean;
-    border?: boolean;
-}
